+++ src/content/panorama/scripts/custom_game/loading_screen/loading_screen_tips.js
@@ -76,6 +76,12 @@ var tipsGeneral = [{
     },{
         img: 'file://{images}/custom_game/hints/hint_discord.png',
         txt: '#hintDiscord'
+    },{
+        img: 'file://{images}/custom_game/hints/hint_bots_difficulty.png',
+        txt: '#hintBotsDifficulty'
+    },{
+        img: 'file://{images}/custom_game/hints/hint_duels.png',
+        txt: '#hintDuels'
     },
 ];